+++ src/components/SpecRenderer.vue
@@ -1,5 +1,23 @@
 <template>
-  <div class="wrapper">
+  <div class="spec-renderer-wrapper">
+    <SlideOut
+      v-if="tableOfContents"
+      class="slideout-toc"
+      :title="parsedDocument?.name || 'Table of Contents'"
+      :visible="slideoutTocVisible"
+      @close="slideoutTocVisible = false"
+    >
+      <SpecRendererToc
+        ref="specRendererSlideoutTocRef"
+        :base-path="basePath"
+        class="spec-renderer-toc"
+        :control-browser-url="controlBrowserUrl"
+        :current-path="currentPath"
+        :table-of-contents="tableOfContents"
+        @item-selected="itemSelected"
+      />
+    </SlideOut>
+
     <aside>
       <SpecRendererToc
         v-if="tableOfContents"
@@ -12,6 +30,22 @@
         @item-selected="itemSelected"
       />
     </aside>
+
+    <!-- small screen menu button - hidden on kui-breakpoint-tablet -->
+    <div
+      v-if="tableOfContents"
+      class="spec-renderer-small-screen-header"
+    >
+      <button
+        class="slideout-toc-trigger-button"
+        type="button"
+        @click="openSlideoutToc"
+      >
+        <MenuIcon class="menu-icon" />
+        Menu
+      </button>
+    </div>
+
     <div class="doc">
       <SpecDocument
         v-if="parsedDocument && currentPath"
@@ -29,10 +63,12 @@
 </template>
 
 <script setup lang="ts">
-import { watch, ref } from 'vue'
+import { watch, ref, nextTick } from 'vue'
 import composables from '../composables'
 import SpecRendererToc from './spec-renderer-toc/SpecRendererToc.vue'
 import SpecDocument from './spec-document/SpecDocument.vue'
+import { MenuIcon } from '@kong/icons'
+import SlideOut from './common/SlideOut.vue'
 
 const props = defineProps({
   /**
@@ -122,6 +158,8 @@ const currentPath = ref<string>(props.currentPath)
 
 const itemSelected = (id: any) => {
   currentPath.value = id
+
+  slideoutTocVisible.value = false
 }
 
 const emit = defineEmits<{
@@ -130,12 +168,28 @@ const emit = defineEmits<{
 
 
 const specRendererTocRef = ref<InstanceType<typeof SpecRendererToc> | null>(null)
+const specRendererSlideoutTocRef = ref<InstanceType<typeof SpecRendererToc> | null>(null)
+const slideoutTocVisible = ref<boolean>(false)
 
 /**
  * re-emits path-not-found event so application that consumes SpecRender component can handle 404
  */
 const relayPathNotFound = (requestedPath: string): void => {
   emit('path-not-found', requestedPath)
+}
+
+const openSlideoutToc = async (): Promise<void> => {
+  slideoutTocVisible.value = true
+
+  await nextTick() // wait for slideout to open
+
+  if (specRendererSlideoutTocRef.value?.$el?.scrollTo) {
+    const scrollPosition = await specRendererSlideoutTocRef.value?.getActiveItemScrollPosition()
+
+    specRendererSlideoutTocRef.value?.$el.scrollTo({
+      top: scrollPosition - 50, // offset 50px so it doesn't stick to the top
+    })
+  }
 }
 
 watch(() => ({
@@ -158,6 +212,7 @@ watch(() => ({
     withCredentials: props.withCredentials,
     currentPath: currentPath.value,
   })
+
   if (props.traceParsing) {
     console.log('parsedDocument:', parsedDocument.value)
     console.log('tableOfContents:', tableOfContents.value)
@@ -173,29 +228,93 @@ watch(specRendererTocRef, async (val) => {
     const scrollPosition = await val.getActiveItemScrollPosition()
 
     val.$el.scrollTo({
-      top: scrollPosition - 50, // offset 50 so it doesn't stick to the top
+      top: scrollPosition - 50, // offset 50px so it doesn't stick to the top
     })
   }
 })
 </script>
 
 <style lang="scss" scoped>
-// TODO: change when implementing generic(responsive) SpecRender layout
-aside {
+.spec-renderer-wrapper {
+  box-sizing: border-box;
   display: flex;
-  flex-shrink: 0;
-  height: 100%;
-  overflow: visible;
-  width: 320px;
-}
-.doc {
-  flex: 1;
-  overflow: visible;
-  padding: var(--kui-space-60, $kui-space-60);
-}
-.wrapper {
-  display: flex;
+  flex-direction: column;
   height: 100vh;
+  position: relative;
+
+  @media (min-width: $kui-breakpoint-tablet) {
+    flex-direction: row;
+  }
+
+  .slideout-toc {
+    :deep(.slideout-container) {
+      padding-left: var(--kui-space-0, $kui-space-0);
+
+      .slideout-content {
+        padding-right: var(--kui-space-0, $kui-space-0);
+      }
+    }
+
+    @media (min-width: $kui-breakpoint-tablet) {
+      display: none;
+    }
+  }
+
+  aside {
+    display: none;
+
+    @media (min-width: $kui-breakpoint-tablet) {
+      display: flex;
+      flex-shrink: 0;
+      height: 100%;
+      overflow: visible;
+      width: 320px;
+    }
+  }
+
+  .spec-renderer-small-screen-header {
+    background-color: var(--kui-color-background-neutral-weakest, $kui-color-background-neutral-weakest);
+    display: flex;
+    padding: var(--kui-space-30, $kui-space-30);
+
+    .slideout-toc-trigger-button {
+      @include default-button-reset;
+
+      align-items: center;
+      color: var(--kui-color-text-neutral, $kui-color-text-neutral);
+      display: flex;
+      font-size: var(--kui-font-size-30, $kui-font-size-30);
+      font-weight: var(--kui-font-weight-medium, $kui-font-weight-medium);
+      gap: var(--kui-space-20, $kui-space-20);
+      line-height: var(--kui-line-height-30, $kui-line-height-30);
+      padding: var(--kui-space-20, $kui-space-20) var(--kui-space-30, $kui-space-30);
+
+      .menu-icon {
+        color: var(--kui-color-text-neutral, $kui-color-text-neutral) !important;
+        height: var(--kui-icon-size-40, $kui-icon-size-40) !important;
+        width: var(--kui-icon-size-40, $kui-icon-size-40) !important;
+      }
+
+      &:hover,
+      &:focus {
+        color: var(--kui-color-text-neutral-strong, $kui-color-text-neutral-strong);
+
+        .menu-icon {
+          color: var(--kui-color-text-neutral-strong, $kui-color-text-neutral-strong) !important;
+        }
+      }
+    }
+
+    @media (min-width: $kui-breakpoint-tablet) {
+      display: none;
+    }
+  }
+
+  .doc {
+    flex: 1;
+    overflow: visible;
+    padding: var(--kui-space-60, $kui-space-60);
+  }
 }
 
 /*
@@ -203,23 +322,33 @@ Styles for SpecRendererToc that need to live here so that they apply to the TOC
 when it's rendered in the context of the SpecRenderer.
 Otherwise host app should have control over these styles.
 */
-.spec-renderer-toc {
+@mixin standalone-spec-renderer-toc($itemPadding: var(--kui-space-70, $kui-space-70)) {
   background-color: var(--kui-color-background, $kui-color-background);
   position: relative; // important, need this for scrolling to selected item
 
   :deep(>) {
     ul > *:first-child {
       // overview item
-      padding: var(--kui-space-70, $kui-space-70);
+      padding: $itemPadding;
       padding-bottom: var(--kui-space-0, $kui-space-0);
     }
   }
 
   :deep(.group-item) {
     &.root {
-      padding-left: var(--kui-space-70, $kui-space-70);
-      padding-right: var(--kui-space-70, $kui-space-70);
+      padding-left: $itemPadding;
+      padding-right: $itemPadding;
     }
+  }
+}
+
+.spec-renderer-toc {
+  @include standalone-spec-renderer-toc;
+}
+
+.slideout-toc {
+  .spec-renderer-toc {
+    @include standalone-spec-renderer-toc($itemPadding: var(--kui-space-40, $kui-space-40));
   }
 }
 </style>